requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.40.0
streamlit>=1.38.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
tenacity>=8.2.0
plotly>=5.18.0